src/frontend/pyproject.toml
[project]
name = "taranis-frontend"
description = "Taranis AI Frontend"
readme = "README.md"
license = { file = "LICENSE.md" }
maintainers = [{ name = "AIT", email = "benjamin.akhras@ait.ac.at" }]

classifiers = [
    "Programming Language :: Python",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.12",
    "Framework :: Flask",
    "Topic :: Security",
]
requires-python = ">=3.12"
dependencies = [
    "Flask",
    "Flask-Caching",
    "Flask-JWT-Extended",
    "flask-htmx",
    "flask-cors",
    "click",
    "pydantic-settings",
    "swagger-ui-py-x",
    "python-dotenv",
    "granian",
    "requests",
    "setuptools",
    "heroicons[jinja]>=2.11.0",
]
dynamic = ["version"]

[tool.setuptools]
packages = ["frontend"]

[build-system]
requires = [
    "setuptools>=45",
    "setuptools_scm[toml]>=6.2",
    "wheel"
]
build-backend = "setuptools.build_meta"

[tool.setuptools_scm]
fallback_version = "0.0.0"
root = "../.."

[project.optional-dependencies]
dev = ["ruff", "pytest", "pytest-flask", "lxml", "build", "wheel", "setuptools_scm", "requests-mock"]

[project.urls]
"Source Code" = "https://github.com/taranis-ai/taranis-ai"

[tool.ruff]
line-length = 142
target-version = "py312"

[tool.ruff.lint.isort]
lines-after-imports = 2

[tool.coverage.run]
branch = true

[tool.pylint.'MESSAGES CONTROL']
max-line-length = 142
disable = """
    missing-module-docstring,
    missing-class-docstring,
    missing-function-docstring
"""

[tool.mypy]
explicit_package_bases = true
namespace-packages = true
exclude = [
    '__main__.py',
    'tests/'
]
